extract base class from Template view to select update mechanism
